influx read/write ready
